+++ frontend/src/app/modules/shared/modules/notification/presentation/notification.component.spec.ts
@@ -43,8 +43,11 @@ describe('NotificationsInboxComponent', () => {
 
   const mapNotificationResponse = (data: NotificationResponse): Notification => {
     const isFromSender = data.channel === 'SENDER';
+    const createdDate = new CalendarDateModel(data.createdDate);
+    const targetDate = new CalendarDateModel(data.targetDate);
     delete data.channel;
-    return { ...data, createdDate: new CalendarDateModel(data.createdDate), isFromSender };
+
+    return { ...data, createdDate, targetDate, isFromSender };
   };
 
   const renderNotificationsInbox = () => {